requirements.txt
appnope==0.1.2
backcall==0.2.0
Brotli==1.0.9
certifi==2021.10.8
chardet==4.0.0
click==8.1.2
dash==1.19.0
dash-bootstrap-components==0.13.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dataclasses==0.6
DateTime==4.3
decorator==5.1.0
entrypoints==0.3
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.8.0
Flask-SSLify==0.1.5
flask-talisman==0.7.0
fpdf2==2.4.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.8.2
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jupyter-client==7.1.0
jupyter-core==4.9.1
MarkupSafe==1.1.1
more-itertools==8.6.0
nest-asyncio==1.5.4
numpy==1.19.5
oauthlib==3.1.1
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
tenacity==8.0.1
tornado==6.1
traitlets==4.3.3
typing_extensions==4.0.1
urllib3==1.25.8
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
zope.interface==5.4.0
